chat-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
chromadb==0.4.21
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==3.17.4
werkzeug==3.0.0
